% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{add_from_file}
\alias{add_from_file}
\title{Read in a file and then add across rows}
\usage{
add_from_file(file)
}
\arguments{
\item{file}{A two column file to read.}
}
\value{
A vector of row sums.
}
\description{
Read in a file and then add across rows
}
\examples{
\dontrun{
add_from_file("some_file_path")
}
}

