% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_wrappers.R
\name{add_wrapper}
\alias{add_wrapper}
\title{Uselessly wrap the add function}
\usage{
add_wrapper(x, y)
}
\arguments{
\item{x}{The first number to add.}

\item{y}{The second number to add.}
}
\value{
The result of computing \code{x + y}.
}
\description{
Uselessly wrap the add function
}
\examples{
add_wrapper(5,3)
add_wrapper(7,7)
}

