% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{divide_from_file}
\alias{divide_from_file}
\title{Read in a file and then divide across rows}
\usage{
divide_from_file(file)
}
\arguments{
\item{file}{A two column file to read.}
}
\value{
A vector of row quotients.
}
\description{
Read in a file and then divide across rows
}
\examples{
\dontrun{
divide_from_file("some_file_path")
}
}

