% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_wrappers.R
\name{divide_wrapper}
\alias{divide_wrapper}
\title{Uselessly wrap the divide function}
\usage{
divide_wrapper(x, y)
}
\arguments{
\item{x}{The number to be divided.}

\item{y}{The number by which to divide..}
}
\value{
The result of computing \code{x / y}.
}
\description{
Uselessly wrap the divide function
}
\examples{
divide_wrapper(5,3)
divide_wrapper(7,7)
}

