% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{multiply_from_file}
\alias{multiply_from_file}
\title{Read in a file and then multiply across rows}
\usage{
multiply_from_file(file)
}
\arguments{
\item{file}{A two column file to read.}
}
\value{
A vector of row products.
}
\description{
Read in a file and then multiply across rows
}
\examples{
\dontrun{
multiply_from_file("some_file_path")
}
}

