% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_wrappers.R
\name{multiply_wrapper}
\alias{multiply_wrapper}
\title{Uselessly wrap the multiply function}
\usage{
multiply_wrapper(x, y)
}
\arguments{
\item{x}{The first number to multiply.}

\item{y}{The second number to multiply.}
}
\value{
The result of computing \code{x * y}.
}
\description{
Uselessly wrap the multiply function
}
\examples{
multiply_wrapper(5,3)
multiply_wrapper(7,7)
}

