% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{subtract_from_file}
\alias{subtract_from_file}
\title{Read in a file and then subtract across rows}
\usage{
subtract_from_file(file)
}
\arguments{
\item{file}{A two column file to read.}
}
\value{
A vector of row subtractions.
}
\description{
Read in a file and then subtract across rows
}
\examples{
\dontrun{
subtract_from_file("some_file_path")
}
}

