% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetic_wrappers.R
\name{subtract_wrapper}
\alias{subtract_wrapper}
\title{Uselessly wrap the subtract function}
\usage{
subtract_wrapper(x, y)
}
\arguments{
\item{x}{The number from which to subtract.}

\item{y}{The subtracted amount.}
}
\value{
The result of computing \code{x - y}.
}
\description{
Uselessly wrap the subtract function
}
\examples{
subtract_wrapper(5,3)
subtract_wrapper(7,7)
}

